+++ actionsheet/src/Actionsheet.tsx
@@ -24,7 +24,7 @@ export function Actionsheet<T>(StyledActionsheet: React.ComponentType<T>) {
         unmountOnExit = true,
         snapPoints = [50],
         // @ts-ignore
-        _experimentalOverlay = true,
+        _experimentalOverlay = false,
         ...props
       }: T & IActionsheetProps,
       ref?: any
@@ -77,20 +77,6 @@ export function Actionsheet<T>(StyledActionsheet: React.ComponentType<T>) {
         finalFocusRef,
         snapPoints,
       ]);
-
-      if (!_experimentalOverlay) {
-        return (
-          <ActionsheetContext.Provider value={contextValue}>
-            <StyledActionsheet
-              ref={ref}
-              style={[StyleSheet.absoluteFill]}
-              {...(props as T)}
-            >
-              {children}
-            </StyledActionsheet>
-          </ActionsheetContext.Provider>
-        );
-      }
 
       return (
         <Overlay