+++ actionsheet/src/index.tsx
@@ -1,18 +1,18 @@
 import { Actionsheet as ActionsheetMain } from './Actionsheet';
 import ActionsheetContent from './ActionsheetContent';
 import { ActionsheetItem } from './ActionsheetItem';
-import { ActionsheetItemText } from './ActionsheetItemText';
 import ActionsheetBackdrop from './ActionsheetBackdrop';
-import ActionsheetDragIndicator from './ActionsheetDragIndicator';
 import { ActionsheetDragIndicatorWrapper } from './ActionsheetDragIndicatorWrapper';
-import { ActionsheetScrollView } from './ActionsheetScrollView';
-import { ActionsheetVirtualizedList } from './ActionsheetVirtualizedList';
-
 import type { IActionsheetComponentType } from './types';
-import { ActionsheetFlatList } from './ActionsheetFlatList';
-import { ActionsheetSectionList } from './ActionsheetSectionList';
-import { ActionsheetSectionHeaderText } from './ActionsheetSectionHeaderText';
-import { ActionsheetIcon } from './ActionsheetIcon';
+// import { ActionsheetItemText } from './ActionsheetItemText';
+// import ActionsheetDragIndicator from './ActionsheetDragIndicator';
+// import { ActionsheetScrollView } from './ActionsheetScrollView';
+// import { ActionsheetVirtualizedList } from './ActionsheetVirtualizedList';
+
+// import { ActionsheetFlatList } from './ActionsheetFlatList';
+// import { ActionsheetSectionList } from './ActionsheetSectionList';
+// import { ActionsheetSectionHeaderText } from './ActionsheetSectionHeaderText';
+// import { ActionsheetIcon } from './ActionsheetIcon';
 
 export function createActionsheet<
   ActionsheetProps,
@@ -63,18 +63,17 @@ export function createActionsheet<
   const Actionsheet = ActionsheetMain(Root) as any;
   Actionsheet.Backdrop = ActionsheetBackdrop(Backdrop, AnimatePresence);
   Actionsheet.Content = ActionsheetContent(Content, AnimatePresence);
-  Actionsheet.DragIndicator = ActionsheetDragIndicator(DragIndicator);
+  Actionsheet.DragIndicator = DragIndicator;
   Actionsheet.DragIndicatorWrapper =
     ActionsheetDragIndicatorWrapper(IndicatorWrapper);
   Actionsheet.Item = ActionsheetItem(Item);
-  Actionsheet.ItemText = ActionsheetItemText(ItemText);
-  Actionsheet.Icon = ActionsheetIcon(Icon);
-  Actionsheet.ScrollView = ActionsheetScrollView(ScrollView);
-  Actionsheet.VirtualizedList = ActionsheetVirtualizedList(VirtualizedList);
-  Actionsheet.FlatList = ActionsheetFlatList(FlatList);
-  Actionsheet.SectionList = ActionsheetSectionList(SectionList);
-  Actionsheet.SectionHeaderText =
-    ActionsheetSectionHeaderText(SectionHeaderText);
+  Actionsheet.ItemText = ItemText;
+  Actionsheet.Icon = Icon;
+  Actionsheet.ScrollView = ScrollView;
+  Actionsheet.VirtualizedList = VirtualizedList;
+  Actionsheet.FlatList = FlatList;
+  Actionsheet.SectionList = SectionList;
+  Actionsheet.SectionHeaderText = SectionHeaderText;
 
   return Actionsheet as IActionsheetComponentType<
     ActionsheetProps,