+++ button/src/ButtonGroup.tsx
@@ -14,6 +14,7 @@ export const ButtonGroup = (
         isAttached,
         isDisabled,
         children,
+        isReversed,
         reversed,
         ...props
       }: any,
@@ -21,7 +22,8 @@ export const ButtonGroup = (
     ) => {
       let computedChildren;
       let childrenArray = React.Children.toArray(flattenChildren(children));
-      childrenArray = reversed ? [...childrenArray].reverse() : childrenArray;
+      childrenArray =
+        isReversed || reversed ? [...childrenArray].reverse() : childrenArray;
 
       if (childrenArray) {
         computedChildren = childrenArray.map((child: any, index: number) => {