+++ hstack/src/index.ts
@@ -1,15 +1,9 @@
-import { HStack as HStackMain } from './HStack';
-
-export function createHStack<StyledHStackProps, StyledHStackSpacerProps>({
+export function createHStack<StyledHStackProps>({
   Root,
-  Spacer,
 }: {
   Root: React.ComponentType<StyledHStackProps>;
-  Spacer: React.ComponentType<StyledHStackSpacerProps>;
 }) {
-  const HStack = HStackMain(Root, Spacer);
+  Root.displayName = 'HStack';
 
-  HStack.displayName = 'HStack';
-
-  return HStack;
+  return Root;
 }