+++ spinner/src/index.tsx
@@ -1,11 +1,13 @@
-import { Spinner } from './Spinner';
-
 export function createSpinner<SpinnerProps>({
   Root,
 }: {
   Root: React.ComponentType<SpinnerProps>;
 }) {
-  const SpinnerTemp = Spinner(Root);
-  SpinnerTemp.displayName = 'Spinner';
-  return SpinnerTemp;
+  Root.displayName = 'Spinner';
+  Root.defaultProps = {
+    // @ts-ignore
+    focusable: true,
+    accessibilityLabel: 'loading',
+  };
+  return Root;
 }