+++ actionsheet/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @gluestack-ui/actionsheet
 
+## 0.2.19
+
+### Patch Changes
+
+- Fixed overlay components to take overlay
+
+## 0.2.18
+
+### Patch Changes
+
+- fix: figma refactor
+
+## 0.2.17
+
+### Patch Changes
+
+- fix: default value of \_experimentalOverlay to false
+
 ## 0.2.16
 
 ### Patch Changes