+++ modal/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @gluestack-ui/modal
 
+## 0.1.20
+
+### Patch Changes
+
+- Fixed overlay components to take overlay
+
+## 0.1.19
+
+### Patch Changes
+
+- fix: default value of \_experimentalOverlay to false
+
 ## 0.1.18
 
 ### Patch Changes