+++ slider/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @gluestack-ui/slider
 
+## 0.1.12
+
+### Patch Changes
+
+- Reversed functionality fixed
+
+## 0.1.11
+
+### Patch Changes
+
+- fix: figma refactor
+
 ## 0.1.10
 
 ### Patch Changes